Start working on AssertFixer
