+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/step/pathes/ComputeClasspath.java
@@ -1,5 +1,6 @@
-package fr.inria.spirals.repairnator.process.step;
+package fr.inria.spirals.repairnator.process.step.pathes;
 
+import fr.inria.spirals.repairnator.process.step.AbstractStep;
 import fr.inria.spirals.repairnator.states.PipelineState;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
 import fr.inria.spirals.repairnator.process.maven.MavenHelper;
