+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/serializer/InspectorTimeSerializer.java
@@ -7,12 +7,12 @@ import fr.inria.spirals.repairnator.Utils;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
 import fr.inria.spirals.repairnator.process.step.BuildProject;
 import fr.inria.spirals.repairnator.process.step.CloneRepository;
-import fr.inria.spirals.repairnator.process.step.ComputeClasspath;
-import fr.inria.spirals.repairnator.process.step.ComputeSourceDir;
+import fr.inria.spirals.repairnator.process.step.pathes.ComputeClasspath;
+import fr.inria.spirals.repairnator.process.step.pathes.ComputeSourceDir;
 import fr.inria.spirals.repairnator.process.step.ResolveDependency;
 import fr.inria.spirals.repairnator.process.step.checkoutrepository.CheckoutBuggyBuild;
 import fr.inria.spirals.repairnator.process.step.gatherinfo.GatherTestInformation;
-import fr.inria.spirals.repairnator.process.step.NopolRepair;
+import fr.inria.spirals.repairnator.process.step.repair.NopolRepair;
 import fr.inria.spirals.repairnator.process.step.push.PushIncriminatedBuild;
 import fr.inria.spirals.repairnator.process.step.TestProject;
 import fr.inria.spirals.repairnator.serializer.engines.SerializedData;
