+++ repairnator/repairnator-pipeline/src/test/java/fr/inria/spirals/repairnator/process/step/TestNopolRepair.java
@@ -2,8 +2,10 @@ package fr.inria.spirals.repairnator.process.step;
 
 import ch.qos.logback.classic.Level;
 import fr.inria.jtravis.entities.Build;
-import fr.inria.jtravis.helpers.BuildHelper;
 import fr.inria.spirals.repairnator.BuildToBeInspected;
+import fr.inria.spirals.repairnator.process.step.pathes.ComputeClasspath;
+import fr.inria.spirals.repairnator.process.step.pathes.ComputeSourceDir;
+import fr.inria.spirals.repairnator.process.step.repair.NopolRepair;
 import fr.inria.spirals.repairnator.states.PipelineState;
 import fr.inria.spirals.repairnator.states.ScannedBuildStatus;
 import fr.inria.spirals.repairnator.Utils;
